# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ambulance & Emergency Warehouse Management System API — Damascus
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /equipment:
    get:
      operationId: listEquipment
      tags: [equipment]
      summary: List equipment with optional filters
      parameters:
        - name: condition
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 5000 }
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, condition, quantity, manufactureYear, createdAt]
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Paginated list of equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentListResponse"
    post:
      operationId: createEquipment
      tags: [equipment]
      summary: Create new equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentInput"
      responses:
        "201":
          description: Created equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/{id}:
    get:
      operationId: getEquipment
      tags: [equipment]
      summary: Get equipment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
        "404":
          description: Not found
    put:
      operationId: updateEquipment
      tags: [equipment]
      summary: Update equipment
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentUpdate"
      responses:
        "200":
          description: Updated equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    delete:
      operationId: deleteEquipment
      tags: [equipment]
      summary: Delete equipment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /equipment/{id}/history:
    get:
      operationId: getEquipmentCardHistory
      tags: [equipment]
      summary: Get an equipment card with linked custodies and movement history
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: type
          in: query
          schema:
            type: string
            enum: [in, init, out, adjust, custody_out, custody_return, damage, central_return]
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: document
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Equipment card, related custodies, and chronological history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentHistoryResponse"
        "404":
          description: Not found
...
  /transactions/out:
    post:
      operationId: createOutTransaction
      tags: [transactions]
      summary: Record an outbound transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutTransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    Equipment:
      type: object
      required: [id, name, condition, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        equipmentType:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        condition:
          type: string
          enum: [good, maintenance, broken, consumed, needs_inspection]
        manufactureYear:
          type: ["integer", "null"]
        originCountry:
          type: ["string", "null"]
        currentHolder:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        quantity:
          type: integer
          default: 1
        minQuantity:
          type: integer
          default: 0
        maintenanceSentAt:
          type: ["string", "null"]
        maintenanceReturnedAt:
          type: ["string", "null"]
        maintenanceNotes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    EquipmentMovement:
      type: object
      required: [id, type, documentNumber]
      properties:
        id: { type: integer }
        type:
          type: string
          enum: [in, out, adjust, custody_out, custody_return, damage, central_return]
        quantity: { type: ["integer", "null"] }
        partyName: { type: ["string", "null"] }
        holderName: { type: ["string", "null"] }
        documentNumber: { type: string }
        documentDate: { type: ["string", "null"], format: date }
        custodyNoteNumber: { type: ["string", "null"] }
        custodyDate: { type: ["string", "null"], format: date }
        custodyLocation: { type: ["string", "null"] }
        reason: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: ["string", "null"], format: date-time }
        operatorName: { type: ["string", "null"] }
    EquipmentHistoryResponse:
      type: object
      required: [equipment, custodies, movements, total]
      properties:
        equipment:
          allOf:
            - $ref: "#/components/schemas/Equipment"
            - type: object
              required: [custodyQuantity, availableQuantity]
              properties:
                custodyQuantity: { type: integer }
                availableQuantity: { type: integer }
        custodies:
          type: array
          items:
            $ref: "#/components/schemas/CustodySummary"
        movements:
          type: array
          items:
            $ref: "#/components/schemas/EquipmentMovement"
        total: { type: integer }
...
    EquipmentInput:
      type: object
      required: [name, condition]
      properties:
        name:
          type: string
        equipmentType:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        condition:
          type: string
        manufactureYear:
          type: ["integer", "null"]
        originCountry:
          type: ["string", "null"]
        currentHolder:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        quantity:
          type: integer
        minQuantity:
          type: integer
        maintenanceSentAt:
          type: ["string", "null"]
        maintenanceReturnedAt:
          type: ["string", "null"]
        maintenanceNotes:
          type: ["string", "null"]
    EquipmentUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        equipmentType:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        condition:
          type: ["string", "null"]
        manufactureYear:
          type: ["integer", "null"]
        originCountry:
          type: ["string", "null"]
        currentHolder:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        quantity:
          type: ["integer", "null"]
        minQuantity:
          type: ["integer", "null"]
        maintenanceSentAt:
          type: ["string", "null"]
        maintenanceReturnedAt:
          type: ["string", "null"]
        maintenanceNotes:
          type: ["string", "null"]
    EquipmentListResponse:
      type: object
      required: [equipment, total, page, limit]
      properties:
        equipment:
          type: array
          items:
            $ref: "#/components/schemas/Equipment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    Transaction:
      type: object
      required: [id, type, itemType, documentNumber, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [in, out, init, adjust, custody_out, custody_return, damage, central_return]
        itemType:
          type: string
          enum: [item, equipment]
        itemId:
          type: ["integer", "null"]
        itemName:
          type: ["string", "null"]
        itemUnit:
          type: ["string", "null"]
        equipmentId:
          type: ["integer", "null"]
        equipmentName:
          type: ["string", "null"]
        quantity:
          type: ["integer", "null"]
        recipientId:
          type: ["integer", "null"]
        recipientName:
          type: ["string", "null"]
        recipientPerson:
          type: ["string", "null"]
        exitReasonId:
          type: ["integer", "null"]
        exitReason:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
        batchNumber:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        supplySource:
          type: ["string", "null"]
          enum: [central_warehouses, null]
        documentDate:
          type: ["string", "null"]
        deliveryNoteNumber:
          type: ["string", "null"]
        deliveryNoteDate:
          type: ["string", "null"]
        internalDeliveryNoteNumber:
          type: ["string", "null"]
        internalDeliveryNoteDate:
          type: ["string", "null"]
        deliveryDestination:
          type: ["string", "null"]
          enum: [administrative_building, ambulance_point, null]
        custodyHolderName:
          type: ["string", "null"]
        custodyNoteNumber:
          type: ["string", "null"]
        custodyDate:
          type: ["string", "null"]
        custodyLocation:
          type: ["string", "null"]
        returnCondition:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        documentNumber:
          type: string
        notes:
          type: ["string", "null"]
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
...
    OutTransactionInput:
      type: object
      required: [itemType, itemId, quantity, recipientId, exitReasonId, internalDeliveryNoteNumber, internalDeliveryNoteDate, deliveryDestination]
      properties:
        itemType:
          type: string
          enum: [item]
        itemId:
          type: integer
        equipmentId:
          type: ["integer", "null"]
        quantity:
          type: integer
          minimum: 1
        recipientId:
          type: integer
        recipientPerson:
          type: ["string", "null"]
        exitReasonId:
          type: ["integer", "null"]
        internalDeliveryNoteNumber:
          type: string
          minLength: 1
        internalDeliveryNoteDate:
          type: string
          format: date
        deliveryDestination:
          type: string
          enum: [administrative_building, ambulance_point]
        notes:
          type: ["string", "null"]
...
    CustodySummary:
      type: object
      required:
        - id
        - equipmentId
        - equipmentName
        - quantity
        - returnedQuantity
        - outstandingQuantity
        - holderName
        - deliveryNoteNumber
        - deliveryDate
        - location
        - status
      properties:
        id:
          type: integer
        equipmentId:
          type: integer
        equipmentName:
          type: string
        serialNumber:
          type: ["string", "null"]
        quantity:
          type: integer
        returnedQuantity:
          type: integer
        outstandingQuantity:
          type: integer
        recipientId:
          type: ["integer", "null"]
        holderName:
          type: string
        deliveryNoteNumber:
          type: string
        deliveryDate:
          type: string
          format: date
        location:
          type: string
        status:
          type: string
          enum: [open, partially_returned, returned, damaged, closed]
...